Day 23 Web Dev
We got to learn about how jQuery actually works. Honestly, I did not know much about Cory but now that I know about it, I am not sure I can go without using it. It is just in credibly helpful so well-made and in general it's just perfect. In my opinion. Honestly I'm in love with it. Of course I'm gonna need some practice but for the most part I feel pretty good using it.