Day 24 of Web Dev
Today I got to learn a lot about how NPM, Node.js, and how programming in the modern day actually works. I will say it's more a kin to writing an essay then actually programming. Because you need to know what sentences to write you need to know what ideas you would like to express. And you will encounter times when you don't know something and you need to look in a book to actually figure out what's the correct way of using this information. And that is quite exciting to me.