Day 25 Part 2 and 26 of Web Dev
I got to finish off the things that I was doing for the last three days. This mainly included understanding how Express actually works and really getting comfortable with the command line. It was difficult however, I do feel like I achieved a lot. Now I learned how to work with EJS. Honestly, this is a completely different way of working but strangely, it feels really familiar as well. I do feel that was a lot more practice I can be really good. But I need to start making my own projects.